requirements.txt
# Web framework
Flask==2.3.3
Werkzeug==2.3.7

# Computer vision and image processing
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1

# Deep learning frameworks
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196

# WebSocket client
websocket-client==1.6.4

# Data processing
pandas==2.0.3

# System utilities
psutil==5.9.5
